# demos/sample-rest-stockapp/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Documentation",
        "version": "1.0",
        "title": "Api Documentation",
        "termsOfService": "urn:tos",
        "contact": {},
        "license": { "name": "Apache 2.0", "url": "http://www.apache.org/licenses/LICENSE-2.0" }
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "paths": {
        "/v1/admin/role": {
            "get": {
                "tags": ["Admin Operations"],
                "summary": "Return a list of all roles registered with the Aerospike cluster.",
                "operationId": "getRoles",
                "produces": ["application/json", "application/msgpack"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/RestClientRole" }
                        }
                    },
                    "403": {
                        "description": "Not authorized to read role information",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": ["Admin Operations"],
                "summary": "Create a role on the Aerospike cluster.",
                "operationId": "createRole",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "rcRole",
                        "description": "rcRole",
                        "required": true,
                        "schema": { "$ref": "#/definitions/RestClientRole" }
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "400": {
                        "description": "Invalid role creation parameters",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to create roles.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "409": {
                        "description": "Role already exists.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/admin/role/{name}": {
            "get": {
                "tags": ["Admin Operations"],
                "summary": "Get information about a specific role.",
                "operationId": "getRole",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the role whose information should be retrieved.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/RestClientRole" }
                    },
                    "403": {
                        "description": "Not authorized to read role information",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified role not found",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            },
            "delete": {
                "tags": ["Admin Operations"],
                "summary": "Remove a specific role from the Aerospike cluster.",
                "operationId": "dropRole",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the role to remove.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "403": {
                        "description": "Not authorized to remove roles",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified role not found",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/admin/role/{name}/privilege": {
            "post": {
                "tags": ["Admin Operations"],
                "summary": "Grant a list of privileges to a specific role.",
                "operationId": "grantPrivileges",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the role to which privileges will be granted.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "privileges",
                        "description": "privileges",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/Privilege" }
                        }
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "400": {
                        "description": "Invalid privilege parameters",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to modfiy roles.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified role not found",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/admin/role/{name}/privilege/delete": {
            "patch": {
                "tags": ["Admin Operations"],
                "summary": "Remove a list of privileges from a specific role.",
                "operationId": "revokePrivileges",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the role from which privileges will be removed.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "privileges",
                        "description": "privileges",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/Privilege" }
                        }
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "400": {
                        "description": "Invalid privilege parameters",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to modfiy roles.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified role not found",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/admin/user": {
            "get": {
                "tags": ["Admin Operations"],
                "summary": "Return a list of information about users.",
                "operationId": "getUsers",
                "produces": ["application/json", "application/msgpack"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "type": "array", "items": { "$ref": "#/definitions/User" } }
                    },
                    "403": {
                        "description": "Not authorized to read user information",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": ["Admin Operations"],
                "summary": "Create a new user.",
                "operationId": "createUser",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "userInfo",
                        "description": "userInfo",
                        "required": true,
                        "schema": { "$ref": "#/definitions/RestClientUserModel" }
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "400": {
                        "description": "Invalid user creation parameters",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to create users.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "409": {
                        "description": "User already exists.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/admin/user/{user}": {
            "get": {
                "tags": ["Admin Operations"],
                "summary": "Return information about a specific user.",
                "operationId": "getUser",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/User" } },
                    "403": {
                        "description": "Not authorized to read user information",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified user not found",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            },
            "delete": {
                "tags": ["Admin Operations"],
                "summary": "Remove a user.",
                "operationId": "dropUser",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "403": {
                        "description": "Not authorized to delete users.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified user not found",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            },
            "patch": {
                "tags": ["Admin Operations"],
                "summary": "Change the password of the specified user.",
                "operationId": "changePassword",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "password",
                        "description": "password",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "403": {
                        "description": "Not authorized to modify users.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified user not found",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "Privilege": {
            "type": "object",
            "required": ["code"],
            "properties": {
                "code": {
                    "type": "string",
                    "enum": [
                        "user-admin",
                        "sys-admin",
                        "data-admin",
                        "read",
                        "read-write",
                        "read-write-udf"
                    ]
                },
                "namespace": {
                    "type": "string",
                    "example": "testNS",
                    "description": "Namespace Scope"
                },
                "set": { "type": "string", "example": "testSet", "description": "setName Scope" }
            },
            "title": "Privilege"
        },
...
        "RestClientError": {
            "type": "object",
            "properties": {
                "inDoubt": {
                    "type": "boolean",
                    "example": false,
                    "description": "A boolean specifying whether it was possible that the operation succeeded. This is only included if true."
                },
                "message": {
                    "type": "string",
                    "example": "Error Message",
                    "description": "A message describing the cause of the error."
                }
            },
            "title": "RestClientError",
            "description": "Error object returned from unsuccesful operations."
        },
...
        "RestClientRole": {
            "type": "object",
            "properties": {
                "name": { "type": "string", "example": "customRole" },
                "privileges": { "type": "array", "items": { "$ref": "#/definitions/Privilege" } }
            },
            "title": "RestClientRole"
        },
...
        "RestClientUserModel": {
            "type": "object",
            "properties": {
                "password": { "type": "string" },
                "roles": { "type": "array", "items": { "type": "string" } },
                "username": { "type": "string" }
            },
            "title": "RestClientUserModel"
        },
        "User": {
            "type": "object",
            "properties": {
                "name": { "type": "string" },
                "roles": { "type": "array", "items": { "type": "string" } }
            },
            "title": "User"
        },
...
    }
}